// frontend/src/app/api/generated/model/ticketStatus.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export enum TicketStatus {
  OPEN = 'OPEN',

  IN_PROGRESS = 'IN_PROGRESS',

  RESOLVED = 'RESOLVED',

  CLOSED = 'CLOSED',
}